backend/server/requirements.txt
Django~=5.1
numpy==1.26.3
requests~=2.32.3
asyncio==3.4.3
httplib2==0.22.0
python-dotenv==1.0.1
pytube==15.0.0
av==12.3.0
multipledispatch==1.0.0
attrs~=24.2.0
pillow~=10.4.0
djoser~=2.2.2
cryptography~=43.0.0
djangorestframework~=3.15.2
django-cors-headers==4.3.1
aiohttp==3.10.5
pytz==2024.1
dailymotion==0.2.5
psycopg2==2.9.9
yt-dlp==2024.8.6
more-itertools==10.2.0
nest-asyncio==1.6.0